+++ plugins/modules/server.py
@@ -10,6 +10,16 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 
+CLOUD_CONTAINER_SERVER_PRODUCT_CODE = [
+    "CLDCON1",
+    "CLDCON2",
+    "CLDCON4",
+    "CLDCON6",
+    "CLDCON8",
+]
+DAEMON_JOB_TYPE = "daemon"
+SCHEDULER_JOB_TYPE = "scheduler"
+
 DOCUMENTATION = """
 ---
 module: server
@@ -176,6 +186,14 @@ from ..module_utils.sitehost import SitehostAPI  # noqa: E402
 
 
 class AnsibleSitehostServer:
+    _SERVER_STATE_MAP = {"On": "started", "Off": "stopped"}
+    # converts module user input to appropriate api call state
+    _API_STATE_MAP = {
+        "restarted": "reboot",
+        "started": "power_on",
+        "stopped": "power_off",
+    }
+
     def __init__(self, module, api):
         self.sh_api = api
         self.module = module
@@ -188,7 +206,7 @@ class AnsibleSitehostServer:
         """deletes a server"""
         server_to_delete = self._get_server_by_name()
         if not server_to_delete:  # server does not exist, so just continue
-            self.module.exit_json( msg="Server does not exist", **self.result )
+            self.module.exit_json(msg="Server does not exist", **self.result)
 
         #  check mode
         if self.module.check_mode:
@@ -215,7 +233,7 @@ class AnsibleSitehostServer:
         self.module.exit_json(**self.result)
 
     def handle_power_status(self):
-        """this handles starting, stopping, and restarting servers"""
+        """Handles starting, stopping, and restarting servers"""
         # check if the server exist
         server_to_change_state = self._get_server_by_name()
         if not server_to_change_state:
@@ -226,44 +244,22 @@ class AnsibleSitehostServer:
 
         requested_server_state = self.module.params.get("state")
 
-        # always restart server when requested
-        if requested_server_state == "restarted":
-            #  check mode
-            if self.module.check_mode:
-                self.module.exit_json(changed=True)
+        body = OrderedDict()
+        body["name"] = self.module.params["name"]
 
-            body = OrderedDict()
-            body["name"] = self.module.params["name"]
-            body["state"] = "reboot"
-
-            restart_result = self.sh_api.api_query(
-                path="/server/change_state.json", method="POST", data=body
-            )
-
-            self.sh_api.wait_for_job(job_id=restart_result["return"]["job_id"])
-
-            self.module.exit_json(
-                changed=True,
-                msg=f"{self.module.params['name']} restarted",
-                server={
-                    "label": server_to_change_state["label"],
-                    "name": self.module.params["name"],
-                    "state": "On",
-                },
-            )
-
-        # otherwise get the current server state to check if task can be skipped
+        # get the current server state to check if task can be skipped
         current_server_state = self.sh_api.api_query(
             path="/server/get_state.json",
             query_params={"name": self.module.params["name"]},
         )["return"]["state"]
 
-        server_state_map = {"On": "started", "Off": "stopped"}
-
         # if server is the requested state already, skip task
-        if server_state_map[current_server_state] == requested_server_state:
+        if (
+            AnsibleSitehostServer._SERVER_STATE_MAP[current_server_state]
+            == requested_server_state
+        ):
             self.result["msg"] = (
-                f"server already {server_state_map[current_server_state]}",
+                f"server already {AnsibleSitehostServer._SERVER_STATE_MAP[current_server_state]}",
             )
             self.result["server"] = {
                 "name": self.module.params["name"],
@@ -276,11 +272,8 @@ class AnsibleSitehostServer:
             self.module.exit_json(changed=True)
 
         # the server state is different from requested state, start/stop server
-        body = OrderedDict()
-        body["name"] = self.module.params.get("name")
-        body["state"] = (
-            "power_on" if requested_server_state == "started" else "power_off"
-        )
+
+        body["state"] = AnsibleSitehostServer._API_STATE_MAP[requested_server_state]
 
         state_change_job = self.sh_api.api_query(
             path="/server/change_state.json", method="POST", data=body
@@ -312,23 +305,35 @@ class AnsibleSitehostServer:
         body["image"] = self.module.params["image"]
         body["params[ipv4]"] = "auto"
 
-        resource = self.sh_api.api_query(
+        api_result = self.sh_api.api_query(
             path="/server/provision.json",
             method="POST",
             data=body,
         )
 
-        if resource:
-            self.sh_api.wait_for_job(
-                job_id=resource["return"]["job_id"], state="Completed"
-            )
-
-        self.result["server"] = self._get_server_by_name(resource["return"]["name"])
-        self.result["server"]["password"] = resource["return"]["password"]
-        self.result["msg"] = (
-            f"server created: {resource['return']['name']},"
-            f" with user: root and password: {resource['return']['password']}"
+        job_type = (
+            SCHEDULER_JOB_TYPE
+            if self.module.params["product_code"] in CLOUD_CONTAINER_SERVER_PRODUCT_CODE
+            else DAEMON_JOB_TYPE
         )
+
+        self.sh_api.wait_for_job(
+            job_id=api_result["return"]["job_id"], job_type=job_type
+        )
+
+        if job_type == DAEMON_JOB_TYPE:  #  Non Cloud Container server provisioned
+            self.result["msg"] = (
+                f"server created: {api_result['return']['name']},"
+                f" with user: root and password: {api_result['return']['password']}"
+            )
+            self.result["server"]["password"] = api_result["return"]["password"]
+
+        else:  #  Cloud Container server provisioned
+            self.result[
+                "msg"
+            ] = f"Cloud Container server {api_result['return']['name']} created"
+
+        self.result["server"] = self._get_server_by_name(api_result["return"]["name"])
         self.result["changed"] = True
 
         self.module.exit_json(**self.result)
@@ -375,7 +380,7 @@ class AnsibleSitehostServer:
 
         server_after_upgrade = self._get_server_by_name()
 
-        self.result["msg"] = f"{server_after_upgrade['name']} sucessfully upgraded"
+        self.result["msg"] = f"{server_after_upgrade['name']} successfully upgraded"
         self.result["server"] = server_after_upgrade
         self.result["changed"] = True
 
@@ -392,20 +397,24 @@ class AnsibleSitehostServer:
             self.module.fail_json(msg="ERROR: no name or label given, exiting")
 
     def _get_server_by_name(self, server_name=None):
-        """return a server by its server name"""
+        """
+        return basic server information by its server name
+
+        :params server_name: select the server to retrive, if not set, then it will get the server specified by the "name" parameter.
+        :returns: information about the server, returns None if server does not exist
+        :rtype: dict or None
+        """
         if server_name is None:
             server_name = self.module.params.get("name")
 
         retrieved_server = self.sh_api.api_query(
             path="/server/get_server.json",
             query_params=OrderedDict({"name": server_name}),
+            skip_status_check=True,
         )
 
         # if server exist, return it
-        if retrieved_server["status"]:
-            return retrieved_server["return"]
-
-        return None  # server does not exist
+        return retrieved_server["return"] if retrieved_server["status"] else None
 
 
 def main():