+++ CHANGELOG.rst
@@ -5,6 +5,19 @@ Sitehost.Cloud Release Notes
 .. contents:: Topics
 
 
+v1.2.0
+======
+
+Minor Changes
+-------------
+
+- sitehost.cloud.server - Now supports provisioning Cloud Container Servers.
+
+New Modules
+-----------
+
+- sitehost.cloud.stack - Manages Cloud Containers
+
 v1.1.0
 ======